addedResetFunction
Added a reset function, which resets the zoom to encompass all retained points. Hopefully this limits how much users have to use the scroll button on their mouse to zoom out. It also resets the limits of the plots to be the max/min of the retained points, so deleted points aren't affecting the plot limits.

Updated the keystroke legend generation to include 'r' for reset.